+++ lightly_studio/src/lightly_studio/examples/example_coco.py
@@ -17,7 +17,7 @@ annotations_json = env.path("EXAMPLES_COCO_JSON_PATH", "/path/to/your/dataset/an
 images_path = env.path("EXAMPLES_COCO_IMAGES_PATH", "/path/to/your/dataset")
 
 # Create a DatasetLoader from a path
-dataset = ls.Dataset.create()
+dataset = ls.ImageDataset.create()
 dataset.add_samples_from_coco(
     annotations_json=annotations_json,
     images_path=images_path,