+++ lightly_studio/src/lightly_studio/models/video.py
@@ -1,6 +1,6 @@
 """This module defines the Video and VideoFrame model for the application."""
 
-from typing import TYPE_CHECKING, List, Optional
+from typing import List, Optional
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict
@@ -10,13 +10,6 @@ from sqlmodel import Field, Relationship, SQLModel
 
 from lightly_studio.models.range import FloatRange, IntRange
 from lightly_studio.models.sample import SampleTable, SampleView
-
-if TYPE_CHECKING:
-    from lightly_studio.models.annotation.annotation_base import (
-        AnnotationBaseTable,
-    )
-else:
-    AnnotationBaseTable = object
 
 
 class VideoBase(SQLModel):
@@ -50,7 +43,9 @@ class VideoTable(VideoBase, table=True):
 
     __tablename__ = "video"
     sample_id: UUID = Field(foreign_key="sample.sample_id", primary_key=True)
-    frames: Mapped[List["VideoFrameTable"]] = Relationship(back_populates="video")
+    frames: Mapped[List["VideoFrameTable"]] = Relationship(
+        sa_relationship_kwargs={"lazy": "select"}, back_populates="video"
+    )
     sample: Mapped["SampleTable"] = Relationship()