+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `VideoDataset` class.
+- Added Captions support for videos.
 - Allow creating tags from all samples matching the current filters when no samples are explicitly selected.
 - Added notebook/Colab support and usage snippet to the docs.
 - Added image similarity search via drag-and-drop, file upload, or clipboard paste.
@@ -21,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrated `DatasetQuery.export()` to `Dataset.export()`.
 - Reduced the package size by using opencv-python-headless.
 - AnnotationLabelTable is now linked to a dataset.
+- `lightly_studio.Dataset` class has been renamed to `lightly_studio.ImageDataset`.
+- Renamed `SampleField` to `ImageSampleField`.
 
 ### Deprecated